management_plane/pyproject.toml
# Management Plane - FastAPI application for encoding pipeline and policy engine

[project]
name = "tupl-management_plane"
version = "0.1.0"
description = "Management Plane for Semantic Security - encoding pipeline and policy enforcement"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "fastmcp<3",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "numpy>=2.1.0",
    "PyYAML>=6.0",
    "python-multipart>=0.0.12",
    "sentence-transformers>=3.0.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "supabase>=2.16.0",
    "chromadb>=0.5.0",
    "onnxruntime>=1.17.0",
    "transformers>=4.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "grpcio>=1.62.0",
    "grpcio-tools>=1.62.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
